fix PDU1 example and add PDU2
With PDU1 format < 240 (peer-to-peer), PDU specific contains the target address. Global (255) can also be
used as target address. Then the parameter group is aimed at all devices. In this case, the PGN is formed
only from PDU format.
With PDU2 format >= 240 (broadcast), PDU format together with the Group Extension in the PDU specific
field forms the PGN of the transmitted parameter group.